OpenAI: No more support vim_codex.
On March 23rd, we will discontinue support for the Codex API.
All customers will have to transition to a different model.
Codex was initially introduced as a free limited beta in 2021, and has maintained that status to date.
Given the advancements of our newest GPT-3.5 models for coding tasks,
we will no longer be supporting Codex and encourage all customers to transition to GPT-3.5-Turbo.

Signed-off-by: Steve Jeong <how2soft@gmail.com>